up: keyframe action

+++ src/ui/pre_render.cc
@@ -180,7 +180,7 @@ namespace qk {
 	bool PreRender::solve(int64_t time) {
 		solveAsyncCall();
 
-		_window->actionCenter()->advance_RT(uint64_t(time) / 10); // advance action
+		_window->actionCenter()->advance_RT(uint32_t(uint64_t(time) / 10)); // advance action
 
 		if ( _tasks.length() ) { // solve task
 			auto i = _tasks.begin(), end = _tasks.end();
